Solution for first compile. PROBLEM: You can't use H3VR code in scripts until it's imported, resulting in errors on first project load. Therefore, the importer itself won't compile either. SOLUTION: Define H3VR_IMPORTED after import. "#if" out all code that touches H3VR code with that define. Use git filters to keep the define out of the repo.
NOTE: I originally wanted to use rsp files for this, but Unity doesn't forward rsp defines to the csproj and otherwise tightly controls every aspect of the csproj, so it's not feasible to just modify it.

NOTE: ALL CONTRIBUTORS must run the included "CONTRIBUTORS-InitConfig.sh" script after cloning the repo for the first time to include .gitconfig in the repository's local config. Git doesn't let you do this automatically because it is a potential security vulnerability.